refactor(control): move find_target_node into placement.rs (CI 500-line file gate)
