+++ crates/orca-control/src/placement.rs
@@ -11,7 +11,9 @@
 
 use std::collections::HashMap;
 
-use crate::state::RegisteredNode;
+use orca_core::config::ServiceConfig;
+
+use crate::state::{AppState, RegisteredNode};
 
 /// Outcome of resolving a placement pin against the registered nodes.
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -95,6 +97,93 @@ pub(crate) fn master_hostname() -> String {
 /// these pins resolve to a local deploy instead of an error.
 pub(crate) fn pin_matches_master(pin: &str) -> bool {
     pin == "master" || pin == "localhost" || pin == "127.0.0.1" || pin == master_hostname()
+}
+
+/// Find a registered node matching the service's placement constraint.
+///
+/// Returns `Ok(None)` when no placement node is set, or when the pin
+/// explicitly names the master (`master`, `localhost`, `127.0.0.1`, or the
+/// master's hostname — the master is not in `registered_nodes`), in which
+/// case the caller deploys locally. Errors when the pin matches more than
+/// one node (#124) — never schedule on an arbitrary winner — or matches
+/// nothing at all: a pin for a vanished/mistyped node must fail the deploy,
+/// not silently land the service on the master.
+pub(crate) async fn find_target_node(
+    state: &AppState,
+    config: &ServiceConfig,
+) -> anyhow::Result<Option<u64>> {
+    let Some(target) = config.placement.as_ref().and_then(|p| p.node.as_ref()) else {
+        return Ok(None);
+    };
+    let nodes = state.registered_nodes.read().await;
+    match resolve_placement(&nodes, target) {
+        PlacementResolution::Node(id) => Ok(Some(id)),
+        PlacementResolution::NoMatch => {
+            if pin_matches_master(target) {
+                return Ok(None);
+            }
+            let drained = drained_matches(&nodes, target);
+            if !drained.is_empty() {
+                tracing::error!(
+                    service = %config.name,
+                    pin = %target,
+                    drained_nodes = ?drained,
+                    "placement pin matches only drained node(s) — refusing to schedule"
+                );
+                anyhow::bail!(
+                    "placement pin {target:?} for service {:?} matches only drained node(s) \
+                     {drained:?} — refusing to deploy; undrain the node or repin the service",
+                    config.name,
+                )
+            }
+            let known: Vec<String> = nodes
+                .values()
+                .map(|n| {
+                    format!(
+                        "{} ({}{})",
+                        n.node_id,
+                        n.address,
+                        n.labels
+                            .get("hostname")
+                            .map(|h| format!(", hostname {h}"))
+                            .unwrap_or_default()
+                    )
+                })
+                .collect();
+            tracing::error!(
+                service = %config.name,
+                pin = %target,
+                known_nodes = ?known,
+                "placement pin matches no registered node and is not the master — refusing to schedule"
+            );
+            anyhow::bail!(
+                "placement pin {target:?} for service {:?} matches no registered node and is \
+                 not the master ({:?}) — refusing to deploy; registered nodes: [{}]",
+                config.name,
+                master_hostname(),
+                known.join(", "),
+            )
+        }
+        PlacementResolution::Ambiguous(ids) => {
+            let described: Vec<String> = ids
+                .iter()
+                .filter_map(|id| nodes.get(id))
+                .map(|n| format!("{} ({})", n.node_id, n.address))
+                .collect();
+            tracing::error!(
+                service = %config.name,
+                pin = %target,
+                nodes = ?described,
+                "placement pin matches multiple nodes — refusing to schedule (#124)"
+            );
+            anyhow::bail!(
+                "placement pin {target:?} for service {:?} matches multiple nodes: {} — \
+                 refusing to schedule; make the pin unique (node id, full address, or hostname)",
+                config.name,
+                described.join(", "),
+            )
+        }
+    }
 }
 
 #[cfg(test)]
